Initial update of Roslyn satellites for 15.5
Bump version number to 2.6
Use specific build number in UpdateResources.proj

NOTE: New and modified strings here are in English because the translations
have  not yet been handed back yet. This accurately matches the most up-to-
date 15.5 closed-source satellites that are currently available. The new
translations will be picked up in a subsequent update.